refactor: generateAnalyticsLink and includedServices to utils
